fix: changed game view permissions (#1060)
* changed game view permissions

* added object-level delete permission

* permissions fixes

* Merge remote-tracking branch 'origin/development' into unified_permission_class

* permission fix

* added test, fixed code climate issues

* added mixins to game viewset

* name change
