+++ aimmo/app_settings.py
@@ -1,7 +1,9 @@
+from functools import wraps
+
 from django.conf import settings
 from django.utils.module_loading import import_string
-from permissions import default_preview_user, DummyPermission
 from django.contrib.auth.models import User
+from rest_framework import permissions
 
 #: URL function for locating the game server, takes one parameter `game`
 GAME_SERVER_URL_FUNCTION = getattr(settings, "AIMMO_GAME_SERVER_URL_FUNCTION", None)
@@ -11,6 +13,23 @@ PREVIEW_USER_AIMMO_DECORATOR = getattr(settings, "PREVIEW_USER_AIMMO_DECORATOR",
 IS_PREVIEW_USER_AIMMO_CLASS = getattr(settings, "IS_PREVIEW_USER_AIMMO_CLASS", None)
 IS_TEACHER_CLASS = getattr(settings, "IS_TEACHER_CLASS", None)
 USERS_FOR_NEW_AIMMO_GAME = getattr(settings, "USERS_FOR_NEW_AIMMO_GAME", None)
+
+
+def default_preview_user(view_func):
+    @wraps(view_func)
+    def wrapped(request, *args, **kwargs):
+        return view_func(request, *args, **kwargs)
+
+    return wrapped
+
+
+class DummyPermission(permissions.BasePermission):
+    """
+    Used to mock general permissions
+    """
+
+    def has_permission(self, request, view):
+        return True
 
 
 def get_aimmo_preview_user_decorator():
