+++ aimmo/permissions.py
@@ -1,15 +1,6 @@
-from functools import wraps
-
 from rest_framework import authentication, permissions
 from django.shortcuts import get_object_or_404
-
-
-def default_preview_user(view_func):
-    @wraps(view_func)
-    def wrapped(request, *args, **kwargs):
-        return view_func(request, *args, **kwargs)
-
-    return wrapped
+from app_settings import IsTeacher, IsPreviewUser
 
 
 class CsrfExemptSessionAuthentication(authentication.SessionAuthentication):
@@ -42,15 +33,6 @@ class GameHasToken(permissions.BasePermission):
             return False
 
 
-class DummyPermission(permissions.BasePermission):
-    """
-    Used to mock general permissions
-    """
-
-    def has_permission(self, request, view):
-        return True
-
-
 class CanUserPlay(permissions.BasePermission):
     """
     Used to verify that an incoming request is made by a user
@@ -62,3 +44,24 @@ class CanUserPlay(permissions.BasePermission):
 
     def has_object_permission(self, request, view, obj):
         return obj.can_user_play(request.user)
+
+
+class CanDeleteGameOrReadOnly(permissions.BasePermission):
+    """
+    Used to verify that an incoming request is made by a user
+    that's authorised to delete or view an AIMMO game
+    """
+
+    def has_permission(self, request, view):
+        return True
+
+    def has_object_permission(self, request, view, obj):
+        if request.method in permissions.SAFE_METHODS:
+            return True
+        else:
+            can_play = CanUserPlay().has_object_permission(request, view, obj)
+            return (
+                IsPreviewUser().has_permission(request, view)
+                and IsTeacher().has_permission(request, view)
+                and can_play
+            )
