+++ version.txt
@@ -1,1 +1,1 @@
-0.22.6
+0.22.7
